agente-reclutamiento-v2/requirements.txt
# Dependencias mínimas para conexión a Gemini
google-genai>=0.3.0
python-dotenv>=1.0.0
